+++ src/selection/SelectionMethod.java
@@ -6,5 +6,5 @@ import java.util.List;
 import java.util.Map;
 
 public interface SelectionMethod {
-    List<Equipment> select(List<Equipment> equipment, Map<String, Integer> selectParams);
+    List<Equipment> select(List<Equipment> equipment, Map<String, Double> selectParams);
 }
